add id user in reception

+++ src/modules/controllers/reception.controllers.js
@@ -1,7 +1,7 @@
 const Receptions = require('../../db/models/receptions/Index');
 
 module.exports.getAllReceptions = (req, res) => {
-  Receptions.find().then(result => {
+  Receptions.find({ idUser: req.query.idUser }).then(result => {
     res.send({ data: result });
   });
 };
@@ -14,12 +14,14 @@ module.exports.createNewReception = (req, res) => {
   && body.hasOwnProperty('nameDoctor') 
   && body.hasOwnProperty('date')
   && body.hasOwnProperty('complaint')
+  && body.hasOwnProperty('idUser')
   && body.nameUser
   && body.nameDoctor
   && body.date
-  && body.complaint) {
+  && body.complaint
+  && body.idUser) {
     reception.save(body).then(() => {
-      Receptions.find().then(result => {
+      Receptions.find({ idUser: body.idUser }).then(result => {
         res.send({ data: result });
       });
     });
@@ -31,9 +33,12 @@ module.exports.createNewReception = (req, res) => {
 module.exports.editReception = (req, res) => {
   const body = req.body;
 
-  if (body.hasOwnProperty('_id') && body._id) {
+  if (body.hasOwnProperty('_id')
+  && body.hasOwnProperty('idUser')
+  && body._id
+  && body.idUser) {
     Receptions.updateOne({ _id: req.body._id }, req.body).then(() => {
-      Receptions.find().then(result => {
+      Receptions.find({ idUser: req.body.idUser }).then(result => {
         res.send({ data: result });
       });
     });
@@ -44,9 +49,12 @@ module.exports.editReception = (req, res) => {
 
 module.exports.deleteReception = async (req, res) => {
   const query = req.query;
-  if (query.hasOwnProperty('_id') && query._id) {
+  if (query.hasOwnProperty('_id')
+  && query.hasOwnProperty('idUser')
+  && query._id
+  && query.idUser) {
     Receptions.deleteOne({ _id: query._id }).then(() => {
-      Receptions.find().then(result => {
+      Receptions.find({ idUser: query.idUser }).then(result => {
         res.send({data: result});
       });
     });
